// 04-IMPL/mejoredu-portal/src/app/follow/smp/indicadores/interfaces/indicator-data.interface.ts
export interface IndicatorDataI {
    id: number
    meta: string;
    indicador: string;
    linea: string;
    metaBienestar: string;
    resultado1: string;
    resultado2: string;
    resultado3: string;
    resultado4: string;
    resultado5: string;
    resultado6: string;
    periodo: string;
}